Commit tests for batch and batchload with different character encodings.
The tests comprise a test script and a number of batch
files which contain strings encoded with different encodings.
The encoding is specified to openr to open a stream,
and then the stream is passed to batch or batchload.

CMUCL, Clisp, ECL, and Clozure CL all pass these tests.
SBCL passes most tests, but there are two failures;
I have not investigated. ABCL should work.
Comments in the code state that GCL ignores encoding
so I do not expect it to work.