Partly revert commit [b08335729] (Typos fixed).
I was criticized, that that commit introduced some
typos, since the rule 'before an vowel comes "an", not "a"'
is not generally true.
Hopefully I fixed all these issues now. (I think, the
commit was not *completly* wrong).

Probably spelling fixes should only be done by native speakers
(and english is not my  mother tongue), sorry for that.