Fix sech/csch/acsch to work better with clisp
Clisp doesn't support denormals, so we need to account for that in the
tests.  Also, clisp has a fairly broken implementation of acosh (can't
evaluate for args larger than 1d35 or so), and asinh (can't evaluate
for arges larger than sqrt(most-positive)).  We work around these
issues by setting the limit to be sqrt(most-positive), and by using
asinh for acosh for large args because acosh(x) = asinh(x) when x is
large enough.