transl: bring errcatch up-to-date
There have been several inconsistencies between how errcatch forms
behave in the interpreter and translator.

A translated errcatch form would not

* bind *MDEBUG* to NIL (this is an ancient bug that was present in
  Macsyma);
* catch rat errors (see commit 672df1e5);
* print error messages or set the error variable when a lisp error
  would occur (see commit 75f4f6a6);
* catch throws to MACSYMA-QUIT or ERRORSW (see commit f11cabfb)

Now an ERRCATCH macro is available that is similar to the classic
ERRSET.  The interpreter and translator now both handle errcatch
forms in the same way via this macro.

No problems with the test suite, share test suite, or rtest_translator.