Fix building PDF manuals for non-english languages.
"category-macros.texi" was not found, as it is in the
parent directory.
I am not really happy with that patch, as it is another
'breakage' for out-of-source builds, but found no better
solution.
Setting TEXINPUTS to set the search path for
"category-macros.texi" or including it not from the parent
directory (../) but from the real directory where it
resides (e.g. $top_srcdir/doc/info/category-macros.texi)
would be better, but I have no idea how to teach GNU
autoconf/automake how to do that. Sorry.

The PDFs for japanese and russian translation do not
compile, but these are other problems (some referenced
files does not exist (eg. Simplification.texi for russian,
should we use the english files in the meantime?) and
probably some font problems).