Add -I option to rule to build maxima.pdf to tell makeinfo to look for files in $(srcdir)/.. . This enables makeinfo to find category-macros.texi.
This commit reverts 5fbcfe, which caused "make" in doc/info to fail,
and adds the -I option for maxima.pdf in common-lang.mk,
which is included in non-English language Makefile.am.

With this change, "make" in doc/info succeeds, and "make pdf" in
doc/info succeeds in es and pt, but fails in pt_BR. "make pdf" succeeds
in de, and fails in ja and ru.