Fix lisp error when evaluating li[3](-.862)
We get a lisp error when evaluating li[3](-.862) because we need to
compute li[3](1), but li3numer doesn't know how to convert 1 to a
float.  Adjust the call (li3numer 1) to be (li3numer (float 1 x))
because at that point, we have the float value x so we can convert 1
to the appropriate float type.