Fix bug #3934: expand(1/(1+%i)^4) => (-4)^(-1) (unsimplified)
This is an ancient bug that was present in Macsyma.

EXPANDEXPT has been constructing MEXPT expressions marked with
the SIMP flag when there is no reason to assume that expression
is actually simplified.  The expression is of the form
((MEXPT SIMP) E -1) where E is the result of EXPONENTIATE-SUM.

(%i1) expand((1+%i)^-2);
(%o1) 1/(2*%i)

(%i2) expand(%);
(%o2) -%i/2

Now we just use INV instead of manually constructing the MEXPT
expression with the bogus SIMP flag.

No problems with the test suite or share test suite.  New tests
have been added to rtest16.