Documentation in pt_BR: change @value{cedilha} to ç in texinfo files.
This allows building the PDF documentation, @value{cedilha} caused strange
errors. In the pt texinfo files, ç is used instead of @value{cedilha} too.

According to Jaime Villate this change is okay.

Maybe the pt_BR translation can be removed and the pt is enough, but
that should other people decide, I do not speak portuguese...