Fix the Maxima version number in lisp-based builds
There is a comment in configure.ac that I added about 7 years ago
that mentions that whitespace in the AC_INIT line should not be
changed.  The whitespace was recently changed for no reason in
commit 1aa4683d.

This problem was reported on the mailing list by Jinsong Zhao in the
thread "correctly display maxima version".