Fix limit(li[n](sin(x)), x, %pi/2)
For an integer n, the above limit was returned as a noun form.
However, it should evaluate to zeta(n).

This is because when computing the limit we compare the arg with 1.
But the limit of sin(x) is 1+zerob.  Call ridofab to get rid of
zeroa/zerob before comparing the arg with 1.

zeta(n) is now returned as the limit, when n is an integer.