Define lucas numbers better and add comments.
I didn't know what Lucas numbers were and the docs didn't really help
except that Maxima can compute them.  So add a sentence to the docs to
give the recursive definition for the Lucas numbers.

Also add some comments in combin.lisp for the functions that compute
the Lucas numbers.  Add one for the Fibonacci numbers as well.

Finally, neatly indent some of the code for these functions.
Otherwise, no functional changes.