Corrected the permissions for the documentation images in the tmp_html folder
If the example images for the documentation are copied using "cp" if the source tree is write-protected
the images will be, too, and cannot be deleted afterwards due to being write-protected.

the "make distcheck" target tests if all temporary files can be deleted, though, and works on a
write-protected source tree.

On my computer this would repair the "make dist" target if there were no test failures.