Fix bug #608: taylor(x^a,[x],0,1) unsimplified
Also bug #545: multivar taylor gives 1^2

This is an ancient bug that was present in Macsyma.

SETUP-MULTIVAR-DISREP gives genvars that are associated with multivars
the DISREP property so they disrep to 1.  These genvars raised to powers
could be visible as unsimplified 1s raised to powers in output:

(%i1) taylor(x^a,[x],0,1);
(%o1) +1^2*x^a

(%i2) taylor((x*y)^a,[x,y],0,1);
(%o2) +1^4*(x*y)^a

(%i3) taylor(x^a+y^a,[x,y],0,1);
(%o3) +(1^2*y^a+1^2*x^a)

Prior to commit 38d4b2fd the internal genvars themselves could be made
visible by using taytorat:

(%i1) taylor(x^a,[x],0,1);
(%o1) +1^2*x^a

(%i2) taytorat(%);
(%o2) (g137^a)^2*x^a

Now we add to PDISREP! the knowledge that 1^x is equal to 1 for any x,
which prevents the construction of factors like 1^2 above.  (PDISREP!
already knows that x^0 is equal to 1 and x^1 is equal to x for any x.)

No problems with the test suite or share test suite.  New tests have
been added to rtest_taylor.