Fix #3967: elliptic_e(5*%pi/4,1) incorrect
elliptic_e(5*%pi/4,1) was returning 2-1/sqrt(2), but
elliptic_e(5*%pi/4,1),numer returned 2.707... or basically
2+1/sqrt(2).

This happens because we were simplifying elliptic_e(5*%pi/4,1) to
sin(5*%pi/4) = -1/sqrt(2).  But we need to reduce the arg 5*%pi/4 by
%pi so we get sin(%pi/4) = 1/sqrt(2).

Add a test for this and that the symbolic and numerical values are
consistent.

Also needed to update the test for elliptic_e(x,1) to include the
reduction of the arg for the sin part.