Fix #3968: zeta(s) for s < 0 can be expressed in terms of zeta(1-s)
zeta(s) can be written in terms of zeta(1-s), so if s < 0, we can
write it in terms of 1-s, a positive number.  The relationship is

  zeta(s) = 2^s*%pi^(s-1)*sin(%pi/2*s)*gamma(1-s)*zeta(1-s))

This makes, for example, zeta(-7/2) return a result in terms of
zeta(9/2), and when you numerically evaluate it, you get the same
answer as you would get for zeta(-3.5).

This is nicer than just returning zeta(-7/2).

Add a couple of tests to rtest_zeta to verify the result.