Add a TR-ABORT function for setting TR-ABORT when translation fails
This provides something to trace or break on, which is useful when
translation silently fails.  This would also be a useful place to put
a throw or signal if we want to immediately bail when translation
fails instead of continuing like we currently do (although I have no
current plans to make that sort of change).

No problems with the test suite, share test suite or rtest_translator.