Remove references to the obsolete A-EXPR property
Under Maclisp this property was used for storing interpreted array
function definitions, but this has been unused since the port to CL.
Only the A-SUBR property is used now.

See also the commit message for commit ce1ea1c1 for more on these
properties.

No problems with the test suite, share test suite or rtest_translator.