Commit package pslq, implementing PSLQ algorithm for finding integer relations among inexact numbers.
Adapted from commit b345584 of https://github.com/robert-dodier/pslq
which is a fork of the package identify by Andrej Vodopivec:
https://github.com/andrejv/identify

This package contains functions pslq_integer_relation, to find integer
relations, and guess_exact_value, which applies pslq_integer_relation to
guess exact values, in terms of radicals, logarithms, exponentials, and
the constant %pi, for inexact (float and bigfloat) numbers.

As pslq contains Maxima code, it is arranged to autoload,
instead of being compiled into the Maxima image.

This commit contains Texinfo documentation in the reference
manual, and a short list of test cases, which is on the list
of share test files, so it is executed by run_testsuite with
share_tests = true or share_tests = only.