Add defun-checked to define functions needing arg checking
Basically split the guts of `defmfun` to `defun-checked` that does the
dirty work of defining two functions: one is the external function
that does argument checking and the other is the internal
implementation function.  `defmfun` basically calls `defun-checked` to
define the functions and just adds the appropriate properties to the
function.

A change:  the implementation function is now `$foo-impl` instead of
`foo-impl`.  We should probably fix that so the impl name is
unchanged.  Or maybe this is a better impl name?