Use built-in pregexp for sregex routines in stringproc
Since pregexp is now built in to maxima, we can delete the (old) copy
in stringproc.  This also means we don't need to load pregexp in
sregex.mac.  Finally, since pregexp is in its own Lisp package, we
need to update sregex.lisp to prefix all uses of pregexp with the
package name.

sregex might be a little faster now too since pregexp is compiled.

Testsuite passes.