Clean up doc for elliptic_eu
Previously, we the html and tex docs didn't actually define
elliptic_eu in terms of integrals.  It just listed the integrals.  Add
elliptic_eu to the integrals to make it clear.

Add a little extra space around dv in the integral.

Make the html and tex formula for the relationship between elliptic_eu
and elliptic_e match the info version.

Make "sn" be roman instead of being in math font in the html version.