Use m4 to define math equations
Add math.m4 which defines the m4 macro m4_mathjax that can be used to
define equations for use with HTML, info, and TeX.

Rename Elliptic.texi to Elliptic.texi.m4 so that we will preprocess
the file with m4 to get the appropriate texinfo commands for math
equations.

Update Makefile.am to add a rule to preprocess *.texi.m4 to *.texi.

Updated the formulas in Elliptic.texi.m4 to use m4_mathjax.