First cut at displaying help via a browser and the html doc
* build-html-index.lisp
Add build-html-index.lisp that reads the html files to get all the
anchors in the multi-page html files.  This outputs the file
maxima-index-html.lisp that contains anchors and html files that will
be used to create an index.

* doc/info/Makefile.am
Updated to run build-html-index.lisp to create maxima-index-html.lisp.

* cl-info.lisp
Update cl-info::load-primary-index to load maxima-index-html.lisp that
sets up *html-index* to have the anchors and html pages.

Update info-exact to display html if requested.

* macdes.lisp
Update to define several variables to be used to configure how to
display the html files:

    * $browser holds the path to the browser to be used
    * $browser_options is a string for any browser options to be used
    * $url_base is the base of the site to be used for getting the pages
    * $describe_uses_html is true if describe displays the results via html

Also adds the function $hdescribe to use a browser to describe one
item.