To workaround the bug
    integrate(x*exp(-x^2)*sin(x),x,minf,inf)
    limit: variable must be a symbol or subscripted symbol; found: sin(x)

I surrounded the call to $limit in gamma-upper-trans (defined in hayat.lisp) with errcatch ($errormsg set to nil). This change allows Maxima to find the correct
value of this definite integral.

But almost surely there is a bug somewhere else that calls gamma-upper-trans with faulty arguments. The real bug should be fixed, but inserting errcatch here is harmless.

Unmark rtest_limit_extra #81 as a known error.

Tested with Clozure CL 1.12.1 & SBCL 2.2.0