Support windows and unix better for html docs
Based on comments from Gunter, on windows we need to do something
like:

cmd "start /max <url>"

Also on Linux, we should use "xdg-open" to open the docs in the user's
default browsker.  Finally, on Mac, we can use "open".

Thus, in init-cl, we set the default for $browser based on the OS.

If we're defaulting to xdg-open (or open), we can't have
$browser_options, so it to "".

Refactor $hdescribe a little and add some more debugging prints so we
can see what we're trying to do.