Fix implementation browser
Ran some experiments on windows using the system command.  For the
browser we can just use "start /max".  $system takes care of calling
cmd to run the command.

Update $hdescribe to use $browser.  However on Unix systems, we need
to quote the URL because the URL contains "#", which would be ignored
by the shell.  The "#" is needed to get the right id for the topic.

Update the manual slightly.