Change $browser to allow a place holder for the URL
$browser should have exactly one ~A (or similar) in it which will be
replaced by the URL.  This simplifies the implementation a bit, but
does require the user to be more careful in specifying a value for
$browser.

The default now is:

windows: "start /max ~a"
darwin: "open '~a'"
linux: "xdg-open '~a'"

I think this handles all typical uses of opening up help in a browser.
The default, of course, is now changed to "firefox '~a'".