Link returned by macro @urlaands shorted to A&S.
The new documatation macro @urlaands inserts a URL to a page of
Abramowitz and Stegun.  The format of the link was changed to
be "A&S" rather than "Abramowitz and Stegun" and the string "eqn"
was dropped to allow links to Chapters, sections, Figures, etc.

"eqn" was added to the macros calls in the texi.m4 as required.