Add some formulas for student t
The pdf and cdf of the Student's t distribution have closed forms so
include them in the text.

We also added an anchor for beta_incomplete_regularized so we can make
a cross-reference to it.

Also corrected a place where we used N(0,1) instead of Normal(0,1)
that is used everywhere else.