Use correct path for htmldir for looking up html files
maxima and maxima-local were looking for the html files in doc/info.
This is incorrect for maxima where the html files are stored in
<path>/doc/html.

To fix this, add an envvar, MAXIMA_HTMLDIR to maxima-local that sets
the html dir to the doc dir.  This basically overrides what maxima
(installed) sets htmldir to.

Update url_base to use *maxima-htmldir* instead of *maxima-infodir*.

When setting the paths on startup, check if MAXIMA_HTMLDIR is set, and
if so, use it to initialize *maxima-htmldir*.

I tested maxima-local to use html for help, which works fine.  Then I
installed maxima to a test dir and did the same.  It also works.