Index was missing items with encoded characters.
The HTML documentation was missing things like "? '".  This is because
the entries were encoded in hex form, and we didn't see them while
searching.  Find and convert more encoded characters.

We also refactored the code to add a new function
PROCESS-ONE-HTML-FILE to handle one file (useful for testing on just
one file).  Also added a new function HANDLE-SPECIAL-CHARS to handle
the encoding of the special characters.

This makes things a bit slower.  Should consider some alternatives.