Fix for opening html files on windows
The problem appears to be that, previously, the url contained
"foo//bar" in the url.  Windows "start" doesn't like that; it should
be just "foo/bar".

Adjust load-primary-index to set up url_base without the trailing "/".

We also fix up an issue where NIL was getting placed in the path when
*maxima-lang-subdir* was NIL.

Tested the fix on linux where only one "/" is seen.  Didn't test on
windows.  I'll wait for the nightly build.