Fix incorrect value for cdf_lognormal for x < 0
Previously cdf_lognormal(x,m,s) returned 1/2 for x < 0.  This is
incorrect since the pdf is 0 for x < 0.

The problem was a misplaced multiplication by unit_step(x).  Add some
parentheses to make it right.

Update the docs  with the correct formula.

Add a test for cdf_lognormal with negative x.