Fix printing of base-strings in maxima-index-html.lisp
build-html-index.lisp was printing strings with *print-readably* set
to T.  For lisps that distinguish base-strings and strings, this can
be an issue because base-strings can't really be printed readably.
Hence set *print-readbaly* to NIL.  We only print strings, so this
shouldn't affect anything else.

Also, when writing "maxima-index-html.lisp", we want to :supersede if
it exists.  We don't care what the previous contents were, or need to
keep any kind of version around.