Correct some info about the default url_base
It said it was the info directory.  That's not true.  It's the
directory that contains the html files.  In a local git checkout, it
is the info directory, but when installed it's the html directory.

Also add a little bit more info why if you use an other url, it has to
contain exactly the same files as in html directory because the index
is built from that.