Use MS edge for displaying html help on Windows
The current setting of `start "URL"` doesn't work to bring up the item
in a browser.  The correct file is used, but the id is ignored.

Instead use `start msedge -url ~A` to display the file.  MS Edge is
the default because it should be available on all recent versions of
Windows.

Update docs to say what the default is, and also suggest how to change
it so that the user can use Firefox or Chrome.  (They'll have to
figure out for themselves if they're using something different.)

No change for other OSes, which seem to be working ok.