Restructure help outputs and support (partially) wxmaxima
First, we've renamed describe_uses_html to output_format_for_help.
This new variable can be set to one of the values 'text, 'html, or
'wxmaxima, with a default of 'text.  It's an error to set it to
anything else.

When output_format_for_help is set, *help-display-function* is
assigned a symbol naming a function that can be used to display the
desired help topics.