Simplify the subsections in distrib
Previously, the section for a random variable had a subsection for an
introduction and one for the functions/variables.  This change removes
these subsections.  Some random variables didn't have intros anyway.

I think it looks better in the html docs to have the introductory
stuff at the top of the page.  Previously, you'd have to click around
to get to the intro section.

This matches more closely what the subsections in the Special
Functions section does.