Use correct hashtable type
In build-html-index were were setting *html-index* to be an equalp
hashtable.  That we want an equal hashtable, which is how *html-index*
is initialized.  So, instead of setting *html-index* to a new
hashtable, just clear the existing table.

This fixes some errors where we were getting duplicated section names
that should have been marked as duplicate.  (Like "Introduction to
Simplification" and "Introduction to simplification".  Note the
capitalization of "simplification".)

There are still some duplicates that need investigation.