Handle fnindex items for html
In the info file, there are entries for items like "Less than" created
using @fnindex.  We weren't searching for these in the html file.
Thus, add a new regexp to find these.

With this change, you can say "?? less than" and get

 0: Less than  (Relational operators)
 1: Less than or equal  (Relational operators)

Now each of these can be found in the html file.

We also clean up `process-one-html-file` by adding an flet to add a
new entry and also warn about duplicates.