Add "(" and ")" to special chars
Texinfo encodes "(" and ")" to the unicode value, so we need to
replace these to find out the topic that can be searched.

Do this in add-entry so that we always handle them.  (Previously, we
only did it for function/var entries.  This is also needed for fnindex
entries like "Fetch documentation (inexact search)".