Handle fnindex entries better
Previously, fnindex entries in the HTML file looked like
"Logical-conjunction-1".  We replaced the dashes with spaces to get an
key of "Logical conjunction 1".  However, "?? logical con" lists
"Logical conjuction <1>", and we converted " <1>" to "-1" before
passing it on.  This works for other entries but not fnindex.

To make it work, we need to put back the dash so that the key is
"Logical conjuction-1".