rtest_integrate: fix up two tests
Due to the recent changes in commit 881c0a626, we no longer
simplify (%e^(a*x^r))^u to %e^(a*u*x^r) for domain:complex since x^r
can be complex even if x and r are real. Therefore, the integral
changes to a different but equally valid form (the difference lies in
discontinuities which drop out upon differentiating).