m4_math automatically adds @math for info
Previously, calls to m4_math typically had @math{foo} for the info
arg.  Now, let m4_math automatically add @math for the info arg.  This
will allow use to define m4_mathdot to place a dot after the arg so it
doesn't have to be done manually.  This is analogous to mref, mrefdot,
and mrefcomma.