In pslq.mac, cut out leading slash on Lisp file name, because pathname_directory always returns something with a trailing slash, at least for the implementations I tried: SBCL, GCL, ECL, CMUCL, Clozure CL, Clisp, and ABCL.
Fixes the error that GCL was running into when trying to load pslq,
which was due to a path name which looked like "foo/share/pslq//pslq.mac".
I believe such paths are OK in the sense that they are POSIX compliant
and also handled successfully by other Lisp implementations, but anyway it
is cleaner this way.