Fix bug #3926: Various limits give UND where they should give IND and bug #3071: limit of expressions with signum not very powerful
BOTH-SIDE finds one-sided limits and compares them to attempt to find
a two-sided limit.  BOTH-SIDE has not been correctly distinguishing
between und and ind; it has (basically) always been returning und when
the one-sided limits are not equal and not both infinities.

Now BOTH-SIDE has an additional check to distinguish between und and
ind.

This caused 13 test suite tests to fail and one test that was expected
to fail to pass.  The 13 tests that failed were actually incorrect:
they had expected results of und when they should have been ind.

No problems with the test suite, share test suite or rtest_translator.
New tests have been added to rtest_limit and the 13 tests that failed
have been changed so that the expected results are ind.