Fix typo and make sure --disable-build-docs works
First, fix typo: "{$makeinfo_old}" should be "${makeinfo_old}".  That
was preventing us from signaling an error.

Second, make sure if --disable-build-docs is set, we don't try to test
for the makeinfo version.  Do this by moving all of the code inside
the test for enable_build_docs being true.