Skip over html files when building html index
Previously we had some hardwired conditions on what files to skip when
building the index.  This included maxima_singlepage.html and anything
that didn't start with "maxima".

Let's make this a little more general and have the functin
build-html-index read from the file "html-exceptions.txt" for a list
of files we want to ignore when creating the index.  The contents of
this file are created in the rule for building maxima-index-html.lisp.

So, we skip maxima_singlepage.html and js_licenses.html.  However, we
also skip over the html file containing the Function and Variable
Index and the Documentation Category index.  These don't have any
links that we need to know, and they're quite large and basically just
wastes time to parse these for entries.