Expunge share package namespaces.
The package more or less works as intended, to expose the Common Lisp symbol package
machinery to the Maxima user, but it is implemented in a way that is likely to cause
trouble, because various internal functions are redefined in order to accommodate
the namespace functionality, so after loading this package, those functions are then
some old version, and any bug fixes or enhanced functionality is lost.
This approach cannot work right, short of updating the namespaces package every
time there is a change to the corresponding internal function.

Also, various failed tests are reported for different Lisp implementations,
and it seems pointless to sort these out, given that the implementation is
not very good to start with, as described above.

I will move the namespaces code to the maxima-packages repo on Github:
https://github.com/maxima-project-on-github/maxima-packages