Create list of html exceptions in common-lang.mk
doc/info/Makefile has a rule to create a file with a list of
exceptions that is used by build-html-index.lisp to skip over unwanted
files.  We forgot to add that to doc/info/common-lang.mk so that the
html files in the language subdirs get the same thing.

This fixes problems with unknown files in es/pt/pt_BR dirs, but
there's now some (unrelated?) error in doc/info/de.