Handle MathJax with old makeinfo
If makeinfo is older than 6.8 (but more than the min required), we
need to include MathJax ourselves.

To do this, modify texi2html.init.in so that configure can substitute
in the JS script needed to load MathJax.

Also, we need to replace @displaymath with something else, so rename
math.m4 to math.m4.in so configure can substitute something in to tell
the m4 macros not to use @displaymath.

Tested with makeinfo 6.5.  Appears to do the right thing.