Print omitted files a bit more neatly
When printing out the files that are omitted, the output looks like

omit ''/path/to/file''

The double single quotes are kind of funny.  Change the print so that
we produce

omit "/path/to/file"

which looks better.

Also slightly reformatted the code because it was really hard to find
the sequence arg to remove-if-not function.

(Do we need the whole path?  It might be better if we just printed the
filename and extension.)