Only search the last files for the indices
Previously, every file was scanned to see if it contained the function
or category indices.  This doubled the time it took to generate the
html index.

Instead, just check the last two files which typically have the two
indices.  If we get this wrong, we end up wasting just a bit of time
scanning files that don't have anything to add to the index.